git-svn-id: https://subversion.assembla.com/svn/cmosher@9262 884d43a4-ffbe-47e9-90ca-fd8c37d7eb32

+++ epple2sdl/src/steppermotor.h
@@ -40,7 +40,7 @@ public:
 	void setMagnet(const unsigned char magnet, const bool on);
 	unsigned char getTrack()
 	{
-		return this->quarterTrack >> 2;
+		return ((unsigned short)(this->quarterTrack)) >> 2;
 	}
 };
